revert: remove OpenAICompatTranslationEngine base class
Base class is premature for only 2 implementations. Restore
OpenAITranslationEngine and OpenRouterTranslationEngine as standalone
classes implementing TranslationEngine directly.
